+++ lib/src/markers/polygon/polygon_marker.dart
@@ -1,5 +1,4 @@
 import '../../components/marker/marker.dart';
-import '../../components/marker/marker_theme.dart';
 import '../../values/coord.dart';
 import '../../values/value.dart';
 import 'polygon_marker_render.dart';
@@ -10,17 +9,14 @@ class GPolygonMarker extends GGraphMarker {
   set close(bool value) => _close.value = value;
 
   GPolygonMarker({
-    String? id,
-    bool visible = true,
+    super.id,
+    super.visible,
+    super.layer,
+    super.hitTestMode,
+    super.theme,
     required List<GCoordinate> coordinates,
     bool close = true,
-    GGraphMarkerTheme? theme,
     super.render = const GPolygonMarkerRender(),
   }) : _close = GValue<bool>(close),
-       super(
-         id: id,
-         visible: visible,
-         keyCoordinates: coordinates,
-         theme: theme,
-       );
+       super(keyCoordinates: coordinates);
 }